+++ Toggly.FeatureManagement.Astro/src/server/toggly-server.ts
@@ -94,6 +94,7 @@ export class TogglyServer implements TogglyClient {
       identity: this.config.identity,
       groups: this.config.groups,
       traits: this.config.claims,
+      claims: this.config.claims,
       entity: entity ?? null,
     };
   }